修改日志文件 位置

+++ src/main/java/com/sjw/file/io/all/helper/FileHelper.java
@@ -11,13 +11,18 @@ import java.io.IOException;
  */
 public class FileHelper {
 
-    private static final String BASE_PATH = "/home/file-io/";
+    private static final String BASE_PATH = System.getProperty("user.dir") + "/";
     private static final String STREAM_FALG = "stream.txt";
     private static final String BUFFER_FALG = "buffer.txt";
     private static final String FILE_CHANNEL_FALG = "fc.txt";
     private static final String MMAP_FALG = "mmap.txt";
+    private static final String RANDOM_FALG = "raf.txt";
 
     public static File getFile(int type, String model) throws IOException {
+        return getFile(type, model);
+    }
+
+    public static File getFile(int type, String model, boolean isNeedNew) throws IOException {
         String path = BASE_PATH;
         if (type == 1) {
             path = path + model + "-" + STREAM_FALG;
@@ -27,14 +32,20 @@ public class FileHelper {
             path = path + model + "-" + FILE_CHANNEL_FALG;
         } else if (type == 4) {
             path = path + model + "-" + MMAP_FALG;
+        } else if (type == 5) {
+            path = path + model + "-" + RANDOM_FALG;
         } else {
             throw new IllegalArgumentException("file io no this type!");
         }
         File file = new File(path);
         if (file.exists()) {
-            file.delete();
+            if (isNeedNew) {
+                file.delete();
+                file.createNewFile();
+            }
+        } else {
+            file.createNewFile();
         }
-        file.createNewFile();
         return file;
     }
